// src/lib/supabase/types-sample.ts
export type Collection = {
  handle: string;
  title: string;
  description: string;
  seo: SEO;
  parentCategoryTree: {
    id: string;
    name: string;
  }[];
  updatedAt: string;
  path: string;
};
// ...
export type SEO = {
  title: string;
  description: string;
};